refactor(Polls): move chatview related classes to its own package
