+++ Aliucord/src/main/java/com/aliucord/coreplugins/polls/chatview/PollChatView.kt
@@ -1,4 +1,4 @@
-package com.aliucord.coreplugins.polls
+package com.aliucord.coreplugins.polls.chatview
 
 import android.annotation.SuppressLint
 import android.content.Context
